Moved the spatial_encoder to src directory - added example to run the spatial encoder
