+++ src/ephysatlas/spatial_encoder/model.py
